// web/src/client/types.gen.ts
export type ProductNodeTreeDto = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Children
     */
    children?: Array<ProductNodeTreeDto>;
    /**
     * Products
     */
    products?: Array<ProductReadTreeDto>;
};
// ...
export type ProductReadTreeDto = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Buy Price
     */
    buy_price: string;
    /**
     * Sell Price
     */
    sell_price: string;
    /**
     * Quantity
     */
    quantity: number;
    /**
     * Created At
     */
    created_at: string;
};